+++ src/components/ui/BurgerMenu/index.ts
@@ -1,0 +1,1 @@
+export * from './BurgerMenu';
